isAvailable function has added error conditional check.
If fingerprint hardware is detected, but no fingerprints have been registered on the device, then the user is passed the error string "Fingerprint authentication not ready". Otherwise the user is passed the error string "Fingerprint authentication not available".

This is to allow the user to distinguish between these different scenarios. The presence of fingerprint hardware dictates a secure device. Also different UI flows and error reporting can be created based upon being able to make this distinction.

Applies to android only.